services/gateway/requirements.txt
Flask==3.0.0
Werkzeug==3.0.0
cryptography==44.0.1
python-dotenv==1.0.1
requests==2.32.3
PyJWT==2.8.0
python-dotenv==1.0.1